Instead of posting to the API with every search input, the projects are loaded initially, and then filtered in the frontend. This functionality is similar to the projectInclude component which is why I extracted it into a service.
