+++ frontend/src/app/shared/components/autocompleter/project-menu-autocomplete/project-menu-autocomplete.template.html
@@ -29,8 +29,9 @@
       class="op-project-list-modal--body spot-container"
     >
       <spot-text-field
+        [disabled]="(loading$ | async)"
         [placeholder]="text.search_placeholder"
-        [(ngModel)]="searchText"
+        [(ngModel)]="searchableProjectListService.searchText"
         [ngModelOptions]="{standalone: true}"
       >
         <span
@@ -39,12 +40,14 @@
         ></span>
       </spot-text-field>
 
-      <ul
-        op-project-list
-        [projects]="projects$ | async"
-        [searchText]="searchText"
-        [multiSelect]="false"
-      ></ul>
+      <ng-container *ngIf="(loading$ | async) === false; else loadingTemplate">
+        <ul
+          op-project-list
+          [projects]="projects$ | async"
+          [searchText]="searchableProjectListService.searchText"
+          [multiSelect]="false"
+        ></ul>
+      </ng-container>
 
       <div class="spot-action-bar">
         <div class="spot-action-bar--right">
@@ -72,7 +75,33 @@
       </div>
     </div>
   </ng-container>
+
+  <ng-template #loadingTemplate>
+    <op-content-loader
+      class="op-project-list-modal--loading"
+      viewBox="0 0 200 140"
+    >
+      <svg:rect x="0" width="100%" height="12" rx="1" />
+
+      <svg:rect x="10" y="16" width="100%" height="12" rx="1" />
+
+      <svg:rect x="10" y="32" width="100%" height="12" rx="1" />
+
+      <svg:rect x="20" y="48" width="100%" height="12" rx="1" />
+
+      <svg:rect x="20" y="64" width="100%" height="12" rx="1" />
+
+      <svg:rect x="0" y="80" width="100%" height="12" rx="1" />
+
+      <svg:rect x="10" y="96" width="100%" height="12" rx="1" />
+
+      <svg:rect x="10" y="112" width="100%" height="12" rx="1" />
+
+      <svg:rect x="0" y="128" width="100%" height="12" rx="1" />
+    </op-content-loader>
+  </ng-template>
 </spot-drop-modal>
+
 
 <!--<div
   class="dropdown dropdown-relative project-search-results"
