+++ frontend/src/app/shared/components/project-include/project-include.component.html
@@ -36,7 +36,7 @@
         [disabled]="(loading$ | async)"
         [placeholder]="text.search_placeholder"
         name="project-include-search"
-        [(ngModel)]="searchText"
+        [(ngModel)]="searchableProjectListService.searchText"
         [ngModelOptions]="{standalone: true}"
         data-qa-selector="project-include-search"
       >
@@ -53,7 +53,7 @@
           [projects]="projects$ | async"
           [selected]="selectedProjects"
           [includeSubprojects]="includeSubprojects$ | async"
-          [searchText]="searchText"
+          [searchText]="searchableProjectListService.searchText"
           [root]="true"
           (update)="selectedProjects = $event"
           data-qa-selector="project-include-list"
@@ -90,39 +90,39 @@
       </div>
     </form>
   </ng-container>
+
+  <ng-template #loadingTemplate>
+    <op-content-loader
+      class="op-project-list-modal--loading"
+      data-qa-selector="op-project-include--loading"
+      viewBox="0 0 200 140"
+    >
+      <svg:rect x="0" width="12" height="12" rx="1" />
+      <svg:rect x="15" width="100%" height="12" rx="1" />
+
+      <svg:rect x="10" y="16" width="12" height="12" rx="1" />
+      <svg:rect x="25" y="16" width="100%" height="12" rx="1" />
+
+      <svg:rect x="10" y="32" width="12" height="12" rx="1" />
+      <svg:rect x="25" y="32" width="100%" height="12" rx="1" />
+
+      <svg:rect x="20" y="48" width="12" height="12" rx="1" />
+      <svg:rect x="35" y="48" width="100%" height="12" rx="1" />
+
+      <svg:rect x="20" y="64" width="12" height="12" rx="1" />
+      <svg:rect x="35" y="64" width="100%" height="12" rx="1" />
+
+      <svg:rect x="0" y="80" width="12" height="12" rx="1" />
+      <svg:rect x="15" y="80" width="100%" height="12" rx="1" />
+
+      <svg:rect x="10" y="96" width="12" height="12" rx="1" />
+      <svg:rect x="25" y="96" width="100%" height="12" rx="1" />
+
+      <svg:rect x="10" y="112" width="12" height="12" rx="1" />
+      <svg:rect x="25" y="112" width="100%" height="12" rx="1" />
+
+      <svg:rect x="0" y="128" width="12" height="12" rx="1" />
+      <svg:rect x="15" y="128" width="100%" height="12" rx="1" />
+    </op-content-loader>
+  </ng-template>
 </spot-drop-modal>
-
-<ng-template #loadingTemplate>
-  <op-content-loader
-    class="op-project-include--loading"
-    data-qa-selector="op-project-include--loading"
-    viewBox="0 0 200 140"
-  >
-    <svg:rect x="0" width="12" height="12" rx="1" />
-    <svg:rect x="15" width="100%" height="12" rx="1" />
-
-    <svg:rect x="10" y="16" width="12" height="12" rx="1" />
-    <svg:rect x="25" y="16" width="100%" height="12" rx="1" />
-
-    <svg:rect x="10" y="32" width="12" height="12" rx="1" />
-    <svg:rect x="25" y="32" width="100%" height="12" rx="1" />
-
-    <svg:rect x="20" y="48" width="12" height="12" rx="1" />
-    <svg:rect x="35" y="48" width="100%" height="12" rx="1" />
-
-    <svg:rect x="20" y="64" width="12" height="12" rx="1" />
-    <svg:rect x="35" y="64" width="100%" height="12" rx="1" />
-
-    <svg:rect x="0" y="80" width="12" height="12" rx="1" />
-    <svg:rect x="15" y="80" width="100%" height="12" rx="1" />
-
-    <svg:rect x="10" y="96" width="12" height="12" rx="1" />
-    <svg:rect x="25" y="96" width="100%" height="12" rx="1" />
-
-    <svg:rect x="10" y="112" width="12" height="12" rx="1" />
-    <svg:rect x="25" y="112" width="100%" height="12" rx="1" />
-
-    <svg:rect x="0" y="128" width="12" height="12" rx="1" />
-    <svg:rect x="15" y="128" width="100%" height="12" rx="1" />
-  </op-content-loader>
-</ng-template>
