+++ frontend/src/app/shared/components/project-include/project-include.component.sass
@@ -1,9 +1,6 @@
 @import '../../../spot/styles/sass/variables'
 
 .op-project-include
-  &--loading
-    padding: $spot-spacing-1
-
   &--include-all
     display: flex
     align-items: center
