+++ frontend/src/global_styles/content/_project_list_modal.sass
@@ -18,3 +18,6 @@
     @media screen and #{$spot-mq-drop-modal-in-context}
       width: 460px
       max-height: 70vh
+
+  &--loading
+    padding: $spot-spacing-1
